Move transformData into each form's generate.js
generatePdf() now takes a deflection object directly and calls
transformData() internally, removing the intermediate data step
from all callers.

Also move dataSchema out of the metadata object into a named
export in each metadata.js.
